Remove old config logic
