+++ packages/react-router-dev/config.ts
@@ -1,8 +1,6 @@
-export type { ReactRouterConfig as Config } from "./config/config";
-
-// todo: move these
 export type {
+  ReactRouterConfig as Config,
   BuildManifest,
   Preset,
   ServerBundlesFunction,
-} from "./vite/config";
+} from "./config/config";
